+++ packages/clients/diplan/src/config.js
@@ -12,6 +12,10 @@ export default {
     248651.73157077, 5227198.20287631, 928366.12236557, 6118661.62507136,
   ],
   diplan: {
+    link: {
+      href: 'https://example.com',
+      icon: '$vuetify.icons.fullscreen-exit',
+    },
     mergeToMultiGeometries: true,
     validateGeoJson: true,
     metaServices: [
@@ -58,6 +62,11 @@ export default {
         url: 'https://geodienste.hamburg.de/HH_WFS_GAGES?service=WFS&request=GetFeature&version=2.0.0',
       },
     ],
+    groupProperties: {
+      defaultGroup: {
+        label: 'Suchbegriff',
+      },
+    },
     minLength: 3,
     waitMs: 300,
   },
@@ -120,7 +129,6 @@ export default {
     ],
   },
   draw: {
-    enableOptions: true,
     lassos: [
       {
         id: flurstuecke,
@@ -129,19 +137,7 @@ export default {
         id: xplanwfs,
       },
     ],
-    measureOptions: {
-      metres: true,
-      kilometres: true,
-      hectares: true,
-    },
-    selectableDrawModes: ['Point', 'LineString', 'Circle', 'Text', 'Polygon'],
     snapTo: [xplanwfs, flurstuecke],
-    textStyle: {
-      font: {
-        size: [10, 20, 30],
-        family: 'Arial',
-      },
-    },
     style: {
       fill: { color: 'rgb(51 117 212 / 50%)' },
       stroke: {
@@ -153,11 +149,6 @@ export default {
         fillColor: 'rgb(51 117 212 / 50%)',
       },
     },
-  },
-  export: {
-    displayComponent: true,
-    showJpg: false,
-    showPdf: false,
   },
   gfi: {
     mode: 'bboxDot',
@@ -177,13 +168,5 @@ export default {
       'plugin/pins/transformedCoordinate',
       'plugin/pins/coordinatesAfterDrag',
     ],
-  },
-  pins: {
-    toZoomLevel: 9,
-    movable: 'drag',
-    appearOnClick: {
-      show: true,
-      atZoomLevel: 0,
-    },
   },
 }