+++ packages/clients/diplan/src/types.ts
@@ -10,9 +10,26 @@ export interface MetaService {
 export type DrawExtension = 'cut' | 'duplicate' | 'merge'
 export type ExtendedDrawMode = Mode | DrawExtension
 
+export type GeoEditingMode =
+  | 'drawPolygon'
+  | 'drawCircle'
+  | 'merge'
+  | 'cut'
+  | 'lasso'
+  | 'edit'
+  | 'translate'
+  | 'delete'
+
+interface LinkConfig {
+  href: string
+  icon: string
+}
+
 export interface DiplanConfiguration {
+  link: LinkConfig
   mergeToMultiGeometries?: boolean
   validateGeoJson?: boolean
+  renderType?: 'iconMenu' | 'independent'
   metaServices?: MetaService[]
 }