+++ packages/plugins/AddressSearch/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unpublished
 
 - Feature: `BKGParameters` has been exposed as root export.
+- Feature: Add new configuration parameter `component` to be able to interchange the UI component of this plugin.
 
 ## 3.0.0